functions/predict_horse_get/requirements.txt
# Function dependencies, for example:
# package>=version
beautifulsoup4==4.8.2
requests==2.23.0
google-cloud-pubsub==1.4.2
google-cloud-storage==1.26.0
